+++ tfd/libraries/javascript.php
@@ -9,13 +9,40 @@
 			'jquery-ui' => 'https://ajax.googleapis.com/ajax/libs/jqueryui/1.8.12/jquery-ui.min.js'
 		);
 		private static $scripts = array();
+		private static $script = array();
+		private static $ready = array();
+		private static $config = array();
 		
 		function echo_scripts(){
-			if(is_array(self::$scripts)){
+			if(!empty(self::$scripts)){
 				foreach(self::$scripts as $s){
 					$scripts .= "{$s}\n";
 				}
 				echo $scripts;
+			}
+			if(!empty(self::$script) || self::$ready){
+				echo "<script>\n";
+			}
+			if(!empty(self::$script)){
+				foreach(self::$script as $s){
+					$script .= "{$s}\n";
+				}
+				echo $script;
+			}
+			if(!empty(self::$ready)){
+				if(self::$config['library'] == 'mootools'){
+					$ready = "window.addEvent('domready',function(){\n";
+				}elseif(self::$config['library'] == 'jquery'){
+					$ready = "\$(document).ready(function(){\n";
+				}
+				foreach(self::$ready as $s){
+					$ready .= "{$s}\n";
+				}
+				$ready .= "});\n";
+				echo $ready;
+			}
+			if(!empty(self::$script) || self::$ready){
+				echo "</script>\n";
 			}
 		}
 		
@@ -26,14 +53,33 @@
 		function library($lib){
 			$src = self::$libraries[$lib];
 			self::$scripts[] = '<script src="'.$src.'"></script>';
+			if($lib == 'mootools'){
+				self::$config['library'] = 'mootools';
+			}elseif($lib == 'jquery'){
+				self::$config['library'] = 'jquery';
+			}
 		}
 		
 		function load($src){
-			self::$scripts[] = '<script src="'.$src.'"></script>';
+			if(is_array($src)){
+				foreach($src as $s){
+					self::$scripts[] = '<script src="'.$s.'"></script>';
+				}
+			}else{
+				self::$scripts[] = '<script src="'.$src.'"></script>';
+			}
 		}
 		
+		// functions, vars, etc. that go outside of the domready function
+		
 		function script($script){
-			self::$scripts[] = $script;
+			self::$script[] = $script;
+		}
+		
+		// function, vars, etc. that go inside of the domready function
+		
+		function ready($script){
+			self::$ready[] = $script;
 		}
 	
 	}